fix resizing bug
the bug was the three.js canvas not wanting to resize height after being increased. this bug is weird and i didnt know a proper way to fix it besides dynamically changing the width/height on window.resize. maybe this is the way, maybe theres a better way. i dont know but it works